chore: move benchmarking logic into the `profiler` crate.
This is another step towards making the benchmarking/profiling logic
distributable in one easy-to-use library. More work to prepare this
library for general use will follow in subsequent PRs.
